FAIR-1901 首屏优化
