+++ src/app/net/login.ts
@@ -9,7 +9,7 @@ import { cryptoRandomInt } from '../../pi/util/math';
 import { wsUrl } from '../config';
 import { AddrInfo, CloudCurrencyType, CurrencyRecord, User, UserInfo, Wallet } from '../store/interface';
 import { Account, getStore, initCloudWallets, LocalCloudWallet, register,setStore } from '../store/memstore';
-import { getWalletTools } from '../utils/commonjsTools';
+import { getCipherTools, getWalletTools } from '../utils/commonjsTools';
 import { CMD } from '../utils/constants';
 import { fetchDeviceId, popNewMessage, popPswBox } from '../utils/tools';
 import { fetchBtcFees, fetchGasPrices, getRealUser, getServerCloudBalance, getUserInfoFromServer, requestAsync, setUserInfo } from './pull';
@@ -135,8 +135,8 @@ export const applyAutoLogin = async () => {
 export const autoLogin = async (conRandom:string) => {
     const deviceId = getStore('setting/deviceId') || await fetchDeviceId();
     console.log('deviceId -------',deviceId);
-    const walletToolsMod = await getWalletTools();
-    const token = walletToolsMod.decrypt(getStore('user/token'),deviceId.toString());
+    const cipherToolsMod = await getCipherTools();
+    const token = cipherToolsMod.decrypt(getStore('user/token'),deviceId.toString());
     const msg = { 
         type: 'wallet/user@auto_login', 
         param: {
